rename: 네이밍 컨벤션, 폴더 구조 변경 - index.tsx가 있는 폴더는 컴포넌트 명과 동일하게, PascalCase로 작성 - index.tsx가 없다면 snake_case로 작성 - use**.ts는 hooks 폴더로 이동 - services 내 파일들은 **Service로 변경
