+++ src/components/BreweryList/BreweryCard/index.tsx
@@ -1,8 +1,8 @@
 import React, { MouseEvent } from "react";
+import Image from "next/image";
+import { useRouter } from "next/navigation";
 import Brewery from "@/types/Brewery";
 import S from "./BreweryCard.module.scss";
-import { useRouter } from "next/navigation";
-import Image from "next/image";
 
 interface Props {
   brewery: Brewery;
@@ -18,7 +18,7 @@ const BreweryCard = ({ brewery, saveSearchInfo }: Props) => {
     city,
     address,
     logo,
-    signatureBeer,
+    beerName,
   } = brewery;
   const fullAddress = `${stateProvince} ${city} ${address}`;
   const router = useRouter();
@@ -46,7 +46,7 @@ const BreweryCard = ({ brewery, saveSearchInfo }: Props) => {
             <span>{breweryIntro}</span>
             <div>
               <span className={S.recommend_title}>추천 맥주</span>
-              <span className={S.recommend_beer}>{signatureBeer.beerName}</span>
+              <span className={S.recommend_beer}>{beerName}</span>
             </div>
           </div>
         </div>
