+++ src/components/brewery-details/contact/index.tsx
@@ -1,19 +1,19 @@
 "use client";
 
 import React from "react";
-import S from "./Contact.module.scss";
-import LocationIcon from "src/components/icons/LocationIcon";
-import PhoneIcon from "src/components/icons/PhoneIcon";
-import LinkIcon from "src/components/icons/LinkIcon";
-import InstagramIcon from "src/components/icons/InstagramIcon";
-import BreweryDetailsProps from "@/types/BreweryDetailsProps";
-import ClockIcon from "src/components/icons/ClockIcon";
-import ChevronIcon from "src/components/icons/ChevronIcon";
-import NaverMaps from "../naver-map/NaverMaps";
 import { NavermapsProvider } from "react-naver-maps";
-import Modal from "@/components/modal/Modal";
-import ModalLink from "@/components/modal/ModalLink";
-import useModal from "@/utils/useModal";
+import BreweryDetailsProps from "@/types/BreweryDetailsProps";
+import S from "./Contact.module.scss";
+import useModal from "@/hooks/useModal";
+import LocationIcon from "@/components/icons/LocationIcon";
+import PhoneIcon from "@/components/icons/PhoneIcon";
+import LinkIcon from "@/components/icons/LinkIcon";
+import InstagramIcon from "@/components/icons/InstagramIcon";
+import ClockIcon from "@/components/icons/ClockIcon";
+import ChevronIcon from "@/components/icons/ChevronIcon";
+import Modal from "@/components/Modal";
+import ModalLink from "@/components/Modal/ModalLink";
+import NaverMaps from "../NaverMaps";
 
 interface Props
   extends Pick<
