+++ src/types/BreweryServiceInterface.ts
@@ -1,9 +1,9 @@
 import Brewery from "./Brewery";
 import BreweryDetailsProps from "./BreweryDetailsProps";
 
-interface BreweryService {
+interface BreweryServiceInterface {
   fetchBreweriesByInputText(query: string): Promise<Brewery[]>;
   fetchBreweryById(breweryId: string): Promise<BreweryDetailsProps>;
 }
 
-export default BreweryService;
+export default BreweryServiceInterface;
